refactor(storage): split forks into dedicated fork module
Move fork operations out of the oversized bucket/forks.ts into a new
lib/fork/ folder (sibling of bucket/ and object/), one file per
operation: list, merge, rebase, plus the internal _legacy list helper.

Public API is unchanged — server.ts re-exports the same symbols, so no
changeset is needed.
